requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
pdfplumber==0.11.4

langchain==0.3.7
langchain-community==0.3.7
faiss-cpu==1.9.0.post1

SQLAlchemy==2.0.35
pymysql==1.1.1
pydantic==2.9.2
python-multipart==0.0.12
